Create nl_NL.php
This is Netherlands Dutch. The file nl_BE.php is very similar, and this one is based on it. The word "wachtwoord" is used for password. Address is spelled as "adres" with one. I've also changed the system_error to something that sounds a little less like a google translate phrase.