Added global password expiration
if the new config parameter 'user_password_valid_for' is set to a date_string like '+1 month', the expiration is active.

there is also a new table column at the 'phpauth_users' table, called 'password_valid_until' which is a timestampt. It get*s updatet by the resetPass() function.

the login() function checks, if the password of a user is expired an returns error = true and also if the password is expired -> $return['password_expired'] = true;